move theme and it's extension to the separate folder
