+++ lib/main.dart
@@ -7,12 +7,12 @@ import 'package:provider/provider.dart';
 import 'package:url_strategy/url_strategy.dart';
 
 import '/utils/utils.dart';
-import 'theme.dart';
 import 'utils/firebase_auth.dart';
 import 'utils/hive_helper.dart';
 import 'utils/notifications/firebase_messaging.dart';
 import 'utils/notifications/local_notifications.dart';
 import 'utils/routes/routes.dart';
+import 'utils/theme/theme.dart';
 import 'utils/theme_helper.dart';
 
 Future<void> main() async {
