+++ lib/screens/announcements/editor/editor_ui.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:provider/provider.dart';
 
-import '../../../theme.dart';
+import '../../../utils/theme/theme.dart';
 import '../../settings/settings_logic.dart';
 import 'editor_logic.dart';
