+++ lib/screens/bottom_bar/bottom_bar_ui.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 
 import '/utils/icons.dart';
-import '../../theme.dart';
+import '../../utils/theme/theme.dart';
 import '../../utils/theme_helper.dart';
 import '../announcements/announcements_screen.dart';
 import '../cabinets_map/cabinets_map_screen.dart';
