+++ lib/screens/login/login_logic.dart
@@ -7,8 +7,8 @@ import 'package:provider/provider.dart';
 import '/screens/settings/settings_logic.dart';
 import '/utils/hive_helper.dart';
 import '../../models/document_model.dart';
-import '../../theme.dart';
 import '../../utils/firebase_auth.dart';
+import '../../utils/theme/theme.dart';
 import '../../utils/theme_helper.dart';
 import '../../utils/utils.dart';
 import 'login_ui.dart';
