+++ lib/screens/schedule/schedule_screen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:r_dotted_line_border/r_dotted_line_border.dart';
 
-import '../../theme.dart';
 import '../../utils/holiday_helper.dart';
+import '../../utils/theme/theme.dart';
 import '../../utils/theme_helper.dart';
 import '../../widgets/snow_widget.dart';
 import '../../widgets/system_bar_cover.dart';
