+++ lib/utils/utils.dart
@@ -10,7 +10,7 @@ import 'package:path_provider/path_provider.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../theme.dart';
+import 'theme/theme.dart';
 import 'theme_helper.dart';
 
 /// Open in browser given [url]
